Update nova_implement_v1.ino
Added functions to store the Setpoint value in the EEPROM. This way, whenever the value changes, it is stored in EEPROM. When the arduino starts, the Setpoint is read from EEPROM.

Added delay of 5s after the initialization. This way, it gives time to arduino to connect to ScadaBR before it starts reading from it.